LeNet-All. First commit
This should be the LeNet-All CNN described in the paper. I had problem running Tensorflow on the computers so it hasn't been tested. The only difference is that I'm not using exactly the same  decay for the learning rate but it's similar.
If this network works then all we have to do is implement the Monte Carlo drop out.